requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Desktop automation
pyautogui==0.9.54
pywinauto==0.6.8
psutil==6.0.0
selenium==4.17.2
webdriver-manager==4.0.1

# Voice - Wake word
pvporcupine==3.0.2

# Voice - STT
openai-whisper==20231117
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.3

# Voice - TTS
pyttsx3==2.90

# LLM - Online APIs
groq==0.4.2
google-generativeai==0.3.2
anthropic==0.8.1

# LLM - Local (Ollama client)
ollama==0.1.6

# Utilities
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1
pillow==10.2.0
pyperclip==1.8.2

# System integration
pywin32==306
pystray==0.19.5

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
